archived old blackjack code
